+++ Flow/src/utopia/flow/caching/multi/TryCache.scala
@@ -1,8 +1,7 @@
 package utopia.flow.caching.multi
 
-import utopia.flow.caching.single.SingleTryCache
-
-import scala.concurrent.duration.FiniteDuration
+import scala.concurrent.ExecutionContext
+import scala.concurrent.duration.{Duration, FiniteDuration}
 import scala.util.Try
 
 /**
@@ -14,38 +13,31 @@ object TryCache
 {
 	/**
 	  * Creates a new cache
-	  * @param failResultDuration The duration which failed results are kept cached
-	  * @param request A function for retrieving a result
-	  * @tparam Key The type of key for this cache
-	  * @tparam Value The result value type on success
+	  * @param failureDuration How long failure results are cached (default = not cached)
+	  * @param successDuration How long success results are cached (default = infinite)
+	  * @param request Function for acquiring a new value
+	  * @param exc Implicit execution context
+	  * @tparam K Type of keys used
+	  * @tparam V Type of values used
 	  * @return A new cache
 	  */
-	def apply[Key, Value](failResultDuration: FiniteDuration)(request: Key => Try[Value]) =
-		MultiCache[Key, Try[Value], SingleTryCache[Value]] { key => SingleTryCache(failResultDuration){ request(key) } }
+	def apply[K, V](failureDuration: FiniteDuration = Duration.Zero, successDuration: Duration = Duration.Inf)
+	               (request: K => Try[V])(implicit exc: ExecutionContext) =
+		ExpiringCache(request) { (_, v) => if (v.isSuccess) successDuration else failureDuration }
 	
 	/**
-	  * Creates a new cache that expires its data
-	  * @param failResultDuration The duration which failed results are kept cached
-	  * @param maxResultDuration The duration which succeeded results are kept cached (should be larger than failResultDuration)
-	  * @param request A function for retrieving a result
-	  * @tparam Key The type of key for this cache
-	  * @tparam Value The result value type on success
+	  * Creates a new cache
+	  * @param failureReferenceDuration How long failure results are strongly referenced
+	  *                                 (default = not strongly referenced)
+	  * @param successReferenceDuration How long success results are strongly referenced (default = infinite)
+	  * @param request Function for acquiring a new value
+	  * @param exc Implicit execution context
+	  * @tparam K Type of keys used
+	  * @tparam V Type of values used
 	  * @return A new cache
 	  */
-	def expiring[Key, Value](failResultDuration: FiniteDuration, maxResultDuration: FiniteDuration)(
-		request: Key => Try[Value]) = ExpiringCache[Key, Try[Value]] {
-		key: Key => SingleTryCache.expiring(failResultDuration, maxResultDuration){ request(key) } }
-	
-	/**
-	  * Creates a new cache that releases its contents after a while, after which it still uses weak references
-	  * @param failureReleaseAfterDuration Duration to keep failure states strongly referenced
-	  * @param successReleaseAfterDuration Duration to keep success states strongly referenced
-	  * @param request A function for retrieving a result
-	  * @tparam Key Type of key for this cache
-	  * @tparam Value Success result type
-	  * @return A new cache
-	  */
-	def releasing[Key, Value](failureReleaseAfterDuration: FiniteDuration, successReleaseAfterDuration: FiniteDuration)
-							 (request: Key => Try[Value]) =
-		ReleasingCache(expiring(failureReleaseAfterDuration, successReleaseAfterDuration)(request))
+	def releasing[K, V <: AnyRef](failureReferenceDuration: FiniteDuration = Duration.Zero,
+	                              successReferenceDuration: Duration = Duration.Inf)
+	                             (request: K => Try[V])(implicit exc: ExecutionContext) =
+		ReleasingCache(request) { (_, v) => if (v.isSuccess) successReferenceDuration else failureReferenceDuration }
 }